refactor: move modules outside of commands
