+++ commands/react_ph/react_ph.py
@@ -1,7 +1,7 @@
 import discord
 import random
 from .consts import POSSIBLE_REACTS, PH_REGEX
-from ..modules.probability import mock_bernoulli
+from modules.probability import mock_bernoulli
 
 
 def is_ph_message(message: discord.Message):
