+++ commands/stats/stats.py
@@ -3,7 +3,7 @@ from discord import app_commands
 
 from bot import TWPlaceClient
 from commands.stats.pat_stats import pat_stats_commands
-from ..modules.supabase import supabaseClient
+from modules.supabase import supabaseClient
 
 
 def register_commands(
